+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#1009](https://github.com/LayerManager/layman/issues/1009) Fill column `description` in `publications` table in prime DB schema. Value is taken from GeoServer for layers and from filesystem for maps.
 ### Changes
 - [#909](https://github.com/LayerManager/layman/issues/909) Upgrade QGIS Server from v3.32.2 to v3.40.2. Also use docker hub repo [layermanager/qgis-server](https://hub.docker.com/repository/docker/layermanager/qgis-server) instead of jirikcz/qgis-server,
+    - Precision error of EPSG:5514 in QGIS WMS GetMap increased from 0.5 m to 3.2 m. The error appears if either data CRS or WMS GetMap CRS is EPSG:5514 and the other one is not.
 - [#1009](https://github.com/LayerManager/layman/issues/1009) PATCH Workspace [Layer](doc/rest.md#patch-workspace-layer)/[Map](doc/rest.md#patch-workspace-map) returns same response as POST Workspace [Layers](doc/rest.md#post-workspace-layers)/[Maps](doc/rest.md#post-workspace-maps) with only `name`, `uuid`, `url` and for Layer also optional `files_to_upload` keys.
 - [#1009](https://github.com/LayerManager/layman/issues/1009) Deprecated endpoint, parameters and keys were removed:
   - key `file_type` was removed from endpoints GET [Publications](doc/rest.md#get-publications)/[Layers](doc/rest.md#get-layers)/[Workspace Layers](doc/rest.md#get-workspace-layers)/[Workspace Layer](doc/rest.md#get-workspace-layer) response